# docs/swagger.yaml
definitions:
  models.Casestudy:
    properties:
      url:
        type: string
    type: object
...
  models.UploadRequest:
    properties:
      file_data:
        type: string
    type: object
  responses.ApplicationResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
  description: Init App aiagent Open Api Spec
  title: Init App aiagent
  version: "1.0"
paths:
...
  /initializ/v1/ai/casestudy:
    get:
      description: Get Case Study Api
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ApplicationResponse'
      summary: Get Case Study
      tags:
      - Case Study Apis
    post:
      description: Save Case Study
      parameters:
      - description: Case Study
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Casestudy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ApplicationResponse'
      summary: Save Case Study
      tags:
      - Case Study Apis
  /initializ/v1/ai/casestudy/{id}:
    delete:
      description: Delete Case Study by ID
      parameters:
      - description: Case Study ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ApplicationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ApplicationResponse'
      summary: Delete Case Study by ID
      tags:
      - Case Study Apis
...
  /initializ/v1/ai/upload:
    post:
      description: Upload Excel File
      parameters:
      - description: File Data in base64 encoded
        in: body
        name: UploadExcel
        required: true
        schema:
          $ref: '#/definitions/models.UploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ApplicationResponse'
      summary: Upload Excel File
      tags:
      - UserData Apis
swagger: "2.0"